Create main
his code sets up a basic XMPP client using the SleekXMPP library. It creates a class called MyClient which inherits from the sleekxmpp.ClientXMPP class. The class has two event handlers, one for "session_start" and another for "message". The "start" method is called when the session starts and it sends the initial presence and gets the roster. The "message" method is called when a message is received, it replys to the sender with the same message.

In the main function, an instance of MyClient is created with a JID and password, and then connected to the XMPP server using the connect method. The process method is then called with the block parameter set to True, which causes the program to run indefinitely and process incoming messages.

This is a basic example, you could add more functionality such as handling file transfer, group messaging, and end-to-end encryption.